Update tsconfig for better web workers support
