--- packages/ui/src/proxyApi/worker/index.d.ts
@@ -1,1 +1,0 @@
-export function workerApi(): Worker
