--- packages/ui/src/proxyApi/worker/index.js
@@ -1,1 +1,0 @@
-export const workerApi = () => new Worker(new URL('./worker', import.meta.url))
