+++ packages/ui/tsconfig.json
@@ -8,6 +8,6 @@
       "@polkadot/types/augment": ["../types/augment/augment-types.ts"]
     }
   },
-  "include": ["dev", "src", "test","./build.js"],
+  "include": ["dev", "src", "test", "./build.js"],
   "exclude": ["src/bounty/**/*", "test/bounty/**/*", "dev/node-mocks/commands/bounty/**/*"]
 }
