+++ tsconfig.json
@@ -3,16 +3,20 @@
     "files": false
   },
   "compilerOptions": {
-    "target": "es2020",
-    "module": "commonjs",
+    "target": "ESNext",
+    "module": "esnext",
+    "moduleResolution": "node",
     "sourceMap": true,
     "strict": true,
-    "esModuleInterop": true,
+    "esModuleInterop": false,
     "resolveJsonModule": true,
     "skipLibCheck": true,
-    "jsx": "react",
-    "lib": ["DOM", "ES2020"],
+    "jsx": "react-jsx",
+    "lib": ["dom", "dom.iterable", "esnext", "webworker"],
     "types": ["jest"],
+    "useUnknownInCatchVariables": false,
+    "allowSyntheticDefaultImports": true,
+    "isolatedModules": true,
     "baseUrl": "."
   },
   "exclude": ["**/node_modules"]
